D3 5688

+++ SW Expert Academy/D3/5688/src/Solution.java
@@ -6,7 +6,7 @@ public class Solution {
 		
 		for(int t=1; t<=T; t++) {
 			long N = sc.nextLong();
-			int num = 1;
+			long num = 1;
 			boolean check = false;
 			while(num*num*num <= N) {
 				if(num*num*num == N) {
